requirements.txt
asttokens==2.0.5
attrs==21.4.0
Automat==20.2.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.4
cssselect==1.1.0
cycler==0.11.0
debugpy==1.6.2
decorator==5.1.1
distlib==0.3.4
entrypoints==0.4
executing==0.8.3
filelock==3.7.1
fonttools==4.34.4
google-api-core==2.8.2
google-auth==2.8.0
google-cloud==0.34.0
google-cloud-bigquery==3.2.0
google-cloud-bigquery-storage==2.13.2
google-cloud-core==2.3.1
google-cloud-storage==2.4.0
google-cloud-vision==2.7.3
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
greenlet==1.1.2
grpcio==1.47.0
grpcio-status==1.47.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipykernel==6.15.1
ipython==8.4.0
itemadapter==0.6.0
itemloaders==1.0.4
jedi==0.18.1
jmespath==1.0.1
jupyter-client==7.3.4
jupyter-core==4.11.1
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.23.0
packaging==21.3
pandas==1.4.3
parsel==1.6.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
playwright==1.23.1
prompt-toolkit==3.0.30
Protego==0.2.1
proto-plus==1.20.6
protobuf==3.20.1
psutil==5.9.1
pure-eval==0.2.2
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyee==8.1.0
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
pyzmq==23.2.0
queuelib==1.6.2
requests==2.28.0
requests-file==1.5.1
rsa==4.8
scipy==1.8.1
Scrapy==2.6.1
scrapy-playwright==0.0.19
seaborn==0.11.2
service-identity==21.1.0
six==1.16.0
stack-data==0.3.0
tldextract==3.3.1
tornado==6.2
traitlets==5.3.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
typing_extensions==4.3.0
urllib3==1.26.9
virtualenv==20.14.1
w3lib==1.22.0
wcwidth==0.2.5
websockets==10.1
zope.interface==5.4.0
